Added new classes for database and feefixing
This is the start to object-orient the rest as well, but so far I did
not have the time to do it. The feefixing now calls any xls. xlsx, csv
and store it in the tables fee_fixing_property or fee_fixing_business
corresponding to the selected radio button in the first screen