problem with FPDF and big files
large PDFs are not displayed and hence I needed to experiment with
other solutions. I installed the latest version of FPDF and so far and
on my machine it works. Need more testing.
I also found a bug in the query for the Bills and Bills register prints
and solved it.